requirements.txt
# Dependencias para OCR de recibos bancarios venezolanos
# Optimizado para recursos limitados

# PyTorch CPU-only (se instala por separado en install.sh)
# torch>=1.9.0
# torchvision>=0.10.0

# OCR y procesamiento de imágenes
easyocr>=1.6.0
opencv-python-headless>=4.5.0
Pillow>=8.0.0
numpy>=1.21.0

# Utilidades
pathlib2>=2.3.0
